Adding context to provider to share the transactions state by the components
