+++ src/App.tsx
@@ -6,6 +6,7 @@ import { GlobalStyle } from "./styles/global";
 import { Header } from './components/Header';
 import { Dashboard } from './components/Dashboard';
 import { NewTransactionModal } from './components/NewTransactionModal';
+import { TransactionsProvider } from './TransactionsContext';
 
 Modal.setAppElement('#root'); // apenas para fins de usabilidade
 
@@ -21,7 +22,7 @@ export function App() {
   }
   
   return (
-    <>
+    <TransactionsProvider>
       <Header onOpenNewTransactionModal={handleOpenNewTransactionModal}/>
       <Dashboard />
 
@@ -31,6 +32,6 @@ export function App() {
             
 
       <GlobalStyle />
-    </>
+    </TransactionsProvider>
   );
 }
