+++ src/components/TransactionsTable/index.tsx
@@ -1,25 +1,12 @@
 
-import { useEffect, useState } from 'react';
-import { api } from '../../services/api';
+import { useContext } from 'react';
+import { TransactionsContext } from '../../TransactionsContext';
 import { Container } from './styles';
 
-interface Transaction {
-    id: number,
-    title: string,
-    amount: number,
-    type: string,
-    category: string,
-    createAt: string,
-}
 
 export function TransactionsTable() {
-    const [transactions, setTransaction] = useState<Transaction[]>([]);
-
-    useEffect(() => {
-        api.get('/transactions')
-        .then(response => setTransaction(response.data.transactions))
-    }, []);
-
+    const transactions = useContext(TransactionsContext);    
+    
     return (
         <Container>
             <table>
